Add files via upload
Simulate tossing a biased coin (a Bernoulli trial) where P[HEAD] = 0.70. (a) Count the number of heads in 50 trials. Record the longest run of heads. (b) Repeat the 50-flip experiment 20, 100, 200, and 1000 times. Use matplotlib to generate a histogram showing the observed number of heads for each case. Comment on the limit of the histogram. (c) Simulate tossing the coin 500 times. Generate a histogram showing the heads run lengths.